Изменен запрос на получение данных из бд

+++ server/index.js
@@ -9,7 +9,8 @@ async function mysqlDb () {
         host: "localhost",
         user: "root",
         database: "dcharts",
-        password: "450979"
+        password: "450979",
+        multipleStatements: true
     });
 }
 mysqlDb();
@@ -19,24 +20,35 @@ const resolvers = {
         getLibrary: async () => {
             try {
                 const [rows] = await connect.execute(`
-                    SELECT library.id,library.name,library.link,
-                    (SELECT GROUP_CONCAT(value,'') FROM value WHERE library_id = library.id) AS value, GROUP_CONCAT(data.value,'') AS data 
-                    FROM library, data WHERE library.id = data.library_id GROUP BY data.library_id
+                    SELECT 
+                        library.id,
+                        library.name,
+                        library.link,
+                        CONCAT(v1.value, ',', v1.label) AS val1,
+                        CONCAT(v2.value, ',', v2.label) AS val2,
+                        GROUP_CONCAT(data.value,'') AS data,
+                        GROUP_CONCAT(data.label,'') AS labels
+                    FROM library
+                    LEFT JOIN value v1 ON v1.library_id = library.id AND v1.label = 'min'
+                    LEFT JOIN value v2 ON v2.library_id = library.id AND v2.label = 'max'
+                    LEFT JOIN data ON data.library_id = library.id
+                    GROUP BY library.id
                 `);
-                // const [rows] = await connect.execute(`
-                //     SELECT library.id, library.name, library.link, value.value, (SELECT GROUP_CONCAT(data.value,'') 
-                //     FROM data WHERE library.id = data.library_id GROUP BY data.library_id) AS data
-                //     FROM library LEFT JOIN value ON library.id = value.library_id
-                // `);
-                
                 let data = []
                 for(let o of rows) {
                     data.push({
                         id: o.id,
                         name: o.name,
                         data: o.data.split(','),
-                        val1: o.value.split(',')[0],
-                        val2: o.value.split(',')[1],
+                        labels: o.labels ? o.labels.split(',') : [],
+                        val1:{
+                            value: o.val1.split(',')[0],
+                            label: o.val1.split(',')[1]
+                        },
+                        val2: {
+                            value: o.val2.split(',')[0],
+                            label: o.val2.split(',')[1]
+                        },
                         link: o.link
                     })
                 }
@@ -57,37 +69,70 @@ const resolvers = {
     },
     Mutation: {
         createNewNote: async (_, {data}) => {
-            console.log('createNewNote', JSON.parse(data))
-            try {
+            // console.log('createNewNote', JSON.parse(data))
+            if ( data != []) {
+                try {
 
-            } catch (e) {
-                console.error(e.message)
+                } catch (e) {
+                    console.error(e.message)
+                }
+                return [JSON.parse(data)]
             }
-            return [JSON.parse(data)]
         },
-        updateNote: async (_, {data}) => {
-            console.log('updateNote', JSON.parse(data))
-            try {
-                
-            } catch (e) {
-                console.error(e.message)
-            }
-            return [JSON.parse(data)];
+        updateNote: async (_, {data}) => {  
+            let notes = JSON.parse(data)
+            if(notes.length != 0) {
+            
+                let updateLibrary = '',
+                    updateData = '',
+                    updateValue = '';
+                // for(let note of notes) {
+                //     updateLibrary += `UPDATE library SET name = ${JSON.stringify(note.name)}, link = ${JSON.stringify(note.link)} WHERE id = ${note.id};`
+                    
+                //     for(let i in note.data) {
+                //         updateData += `UPDATE data SET value = ${note.data[i]} WHERE library_id = ${note.id};`
+                //     }
+                //     updateValue += `UPDATE value v1 JOIN value v2 ON v1.library_id = ${note.id} AND v2.library_id = ${note.id} SET v1.value = ${note.val1} v2.value = ${note.val2};`
+                //     // updateValue += `UPDATE value SET value = ${note.val1} WHERE library_id = ${note.id}`
+                // }
+                // console.log('!!!',updateLibrary)
+                // console.log('!!!!!!',updateData)
+                // console.log('!!!!!!!!!!!',updateValue)
+                //let query = updateLibrary + updateData + updateValue
+                try {
+                    for(let note of notes) {
+                        updateLibrary = `UPDATE library SET name = ${JSON.stringify(note.name)}, link = ${JSON.stringify(note.link)} WHERE id = ${note.id};`
+                        await connect.execute(`${updateLibrary}`);
+
+                        for(let i in note.data) {
+                            updateData = `INSERT INTO data SET value = ${note.data[i]}, `
+                            await connect.execute(`${updateData}`);
+                        }
+                        updateValue = `UPDATE value v1 JOIN value v2 ON v1.library_id = ${note.id} AND v2.library_id = ${note.id} SET v1.value = ${note.val1} v2.value = ${note.val2};`
+                        await connect.execute(`${updateValue}`);
+                    }
+                    //await connect.execute(query);
+                    // await connect.execute(`${updateLibrary}`);
+                    // await connect.execute(`${updateData}`);
+                    // await connect.execute(`${updateValue}`);
+                } catch (e) {
+                    console.error(e.message)
+                }
+                return [JSON.parse(data)]; 
+            }         
         },
         deleteNote: async (_, {data}) => {
-            let noteId = JSON.parse(data).map(e => JSON.parse(e))
-            console.log(noteId)
-            try {
-                await connect.execute(`DELETE FROM library WHERE id in (${noteId})`);
-            } catch (e) {
-                console.error(e.message)
-            }
-            return [JSON.parse(data)];
+            let d = JSON.parse(data)
+            if (d.length != 0) {
+                let noteId = d.map(e => JSON.parse(e))        
+                try {
+                    await connect.execute(`DELETE FROM library WHERE id in (${noteId})`);
+                } catch (e) {
+                    console.error(e.message)
+                }
+                return [JSON.parse(data)];
+            }   
         }
-        // changeDatabase: async (_, {data}) => {
-        //     console.log(JSON.parse(data))
-        //     // return {data}
-        // }
     },
     Library:{
         children:() => {
